requirements.txt
flask==3.0.3
selenium==4.21.0
requests==2.32.3
email-validator==2.1.1
python-dotenv==1.0.1
werkzeug==3.0.3
